Reducing impact of successive identical auth failure
This patch modifies the anti-brute-force mechanism to increase the number of failed counts by only 0.1 when identical
login+password are repeated. If different logins are attempted, the failed counts will be increased by 1 as before.

This is to circumvent the multiple repeated attempts that browsers may perform without user interaction. For example, Firefox tends to automatically retry multiple times previous credentials, leading to triggering evasion when the user did not actually type multiple user/password combinations.

In order to still protect from brute-force, we create a hash of the previous credentials that lead to a failure (see ngx_hash_http_auth_to_uint32()).
If the successive failed attempts are detected to be submitted with the exact same user/password combination, we only increase the failcount by 0.1.